Make AuxKernelBase for Kokkos AuxKernel implementation #30655
