+++ framework/include/auxkernels/AuxKernel.h
@@ -9,27 +9,8 @@
 
 #pragma once
 
-#include "MooseObject.h"
-#include "MooseVariableFE.h"
-#include "SetupInterface.h"
-#include "CoupleableMooseVariableDependencyIntermediateInterface.h"
-#include "MaterialPropertyInterface.h"
-#include "FunctionInterface.h"
-#include "UserObjectInterface.h"
-#include "TransientInterface.h"
-#include "PostprocessorInterface.h"
-#include "DependencyResolverInterface.h"
-#include "RandomInterface.h"
-#include "GeometricSearchInterface.h"
-#include "BlockRestrictable.h"
-#include "BoundaryRestrictable.h"
-#include "Restartable.h"
-#include "MeshChangedInterface.h"
-#include "VectorPostprocessorInterface.h"
+#include "AuxKernelBase.h"
 #include "MooseVariableInterface.h"
-#include "ElementIDInterface.h"
-#include "UserObject.h"
-#include "NonADFunctorInterface.h"
 
 // forward declarations
 template <typename ComputeValueType>
@@ -39,35 +20,11 @@ typedef AuxKernelTempl<Real> AuxKernel;
 typedef AuxKernelTempl<RealVectorValue> VectorAuxKernel;
 typedef AuxKernelTempl<RealEigenVector> ArrayAuxKernel;
 
-class SubProblem;
-class AuxiliarySystem;
-class SystemBase;
-class MooseMesh;
-
 /**
  * Base class for creating new auxiliary kernels and auxiliary boundary conditions.
- *
  */
 template <typename ComputeValueType>
-class AuxKernelTempl : public MooseObject,
-                       public MooseVariableInterface<ComputeValueType>,
-                       public BlockRestrictable,
-                       public BoundaryRestrictable,
-                       public SetupInterface,
-                       public CoupleableMooseVariableDependencyIntermediateInterface,
-                       public FunctionInterface,
-                       public UserObjectInterface,
-                       public TransientInterface,
-                       public MaterialPropertyInterface,
-                       public PostprocessorInterface,
-                       public DependencyResolverInterface,
-                       public RandomInterface,
-                       public GeometricSearchInterface,
-                       public Restartable,
-                       public MeshChangedInterface,
-                       protected VectorPostprocessorInterface,
-                       public ElementIDInterface,
-                       protected NonADFunctorInterface
+class AuxKernelTempl : public AuxKernelBase, public MooseVariableInterface<ComputeValueType>
 {
 public:
   static InputParameters validParams();
@@ -77,7 +34,7 @@ public:
   /**
    * Computes the value and stores it in the solution vector
    */
-  virtual void compute();
+  virtual void compute() override;
 
   /**
    * Nodal or elemental kernel?
@@ -95,14 +52,6 @@ public:
    * @return reference to a variable this kernel is action on
    */
   MooseVariableField<ComputeValueType> & variable() { return _var; }
-
-  const std::set<UserObjectName> & getDependObjects() const { return _depend_uo; }
-
-  void coupledCallback(const std::string & var_name, bool is_old) const override;
-
-  virtual const std::set<std::string> & getRequestedItems() override;
-
-  virtual const std::set<std::string> & getSuppliedItems() override;
 
   /**
    * Override functions from MaterialPropertyInterface for error checking
@@ -150,44 +99,14 @@ protected:
    */
   const typename OutputTools<ComputeValueType>::VariableValue & uOlder() const;
 
-  /**
-   * Whether or not to check for repeated element sides on the sideset to which
-   * the auxkernel is restricted (if boundary restricted _and_ elemental). Setting
-   * this to false will allow an element with more than one face on the boundary
-   * to which it is restricted allow contribution to the element's value(s). This
-   * flag allows auxkernels that evaluate boundary-restricted elemental auxvariables
-   * to have more than one element face on the boundary of interest.
-   */
-  const bool & _check_boundary_restricted;
-
-  /// Subproblem this kernel is part of
-  SubProblem & _subproblem;
-  /// System this kernel is part of
-  SystemBase & _sys;
-  SystemBase & _nl_sys;
-  AuxiliarySystem & _aux_sys;
-
-  /// Thread ID
-  THREAD_ID _tid;
-
-  /// This is a regular kernel so we cast to a regular MooseVariable
+  /// This is a regular kernel so we cast to a regular MooseVariable, hides base _var
   MooseVariableField<ComputeValueType> & _var;
 
   /// Flag indicating if the AuxKernel is nodal
-  bool _nodal;
+  const bool _nodal;
 
   /// Holds the solution at current quadrature points
   const typename OutputTools<ComputeValueType>::VariableValue & _u;
-
-  /// Assembly class
-  Assembly & _assembly;
-
-  /// true if the kernel is boundary kernel, false if it is interior kernels
-  bool _bnd;
-  /// Mesh this kernel is active on
-  MooseMesh & _mesh;
-  /// Dimension of the problem being solved
-  //  unsigned int _dim;
 
   /// Holds the the test functions
   const typename OutputTools<ComputeValueType>::VariableTestValue & _test;
@@ -222,10 +141,6 @@ protected:
   /// The current lower dimensional element
   const Elem * const & _current_lower_d_elem;
 
-  /// Whether we are computing for a lower dimensional variable using boundary restriction, e.g. a
-  /// variable whose block restriction is coincident with a higher-dimensional boundary face
-  const bool _coincident_lower_d_calc;
-
   /// Quadrature point index
   unsigned int _qp;
 
@@ -244,11 +159,6 @@ protected:
   using MooseVariableInterface<ComputeValueType>::mooseVariableBase;
 
 private:
-  void addPostprocessorDependencyHelper(const PostprocessorName & name) const override final;
-  void addUserObjectDependencyHelper(const UserObject & uo) const override final;
-  void
-  addVectorPostprocessorDependencyHelper(const VectorPostprocessorName & name) const override final;
-
   /**
    * Currently only used when the auxiliary variable is a finite volume variable, this helps call
    * through to the variable's \p setDofValue method. This helper is necessary because \p
@@ -257,13 +167,6 @@ private:
    * ComputeValueType is \p RealVectorValue
    */
   void setDofValueHelper(const ComputeValueType & dof_value);
-
-  /// Depend AuxKernelTempls
-  mutable std::set<std::string> _depend_vars;
-  std::set<std::string> _supplied_vars;
-
-  /// Depend UserObjects
-  mutable std::set<UserObjectName> _depend_uo;
 };
 
 template <typename ComputeValueType>
