+++ framework/include/interfaces/NonADFunctorInterface.h
@@ -23,6 +23,14 @@ public:
 
   NonADFunctorInterface(const MooseObject * moose_object);
 
+#ifdef MOOSE_KOKKOS_ENABLED
+  /**
+   * Special constructor used for Kokkos functor copy during parallel dispatch
+   */
+  NonADFunctorInterface(const NonADFunctorInterface & object,
+                        const Moose::Kokkos::FunctorCopy & key);
+#endif
+
 private:
   virtual bool isADObject() const override { return false; }
 };
