+++ framework/src/auxkernels/AuxKernel.C
@@ -26,63 +26,20 @@ template <typename ComputeValueType>
 InputParameters
 AuxKernelTempl<ComputeValueType>::validParams()
 {
-  InputParameters params = MooseObject::validParams();
-  params += BlockRestrictable::validParams();
-  params += BoundaryRestrictable::validParams();
-  params += RandomInterface::validParams();
-  params += MeshChangedInterface::validParams();
-  params += MaterialPropertyInterface::validParams();
-  params += FunctorInterface::validParams();
-  params += GeometricSearchInterface::validParams();
+  InputParameters params = AuxKernelBase::validParams();
 
-  // Add the SetupInterface parameter 'execute_on' with 'linear' and 'timestep_end'
-  params += SetupInterface::validParams();
-  ExecFlagEnum & exec_enum = params.set<ExecFlagEnum>("execute_on", true);
-  exec_enum.addAvailableFlags(EXEC_PRE_DISPLACE);
-  exec_enum = {EXEC_LINEAR, EXEC_TIMESTEP_END};
-  params.setDocString("execute_on", exec_enum.getDocString());
-
-  params.addRequiredParam<AuxVariableName>("variable",
-                                           "The name of the variable that this object applies to");
-
-  params.addParam<bool>("use_displaced_mesh",
-                        false,
-                        "Whether or not this object should use the "
-                        "displaced mesh for computation.  Note that "
-                        "in the case this is true but no "
-                        "displacements are provided in the Mesh block "
-                        "the undisplaced mesh will still be used.");
-  params.addParamNamesToGroup("use_displaced_mesh", "Advanced");
-  params.addParam<bool>("check_boundary_restricted",
-                        true,
-                        "Whether to check for multiple element sides on the boundary "
-                        "in the case of a boundary restricted, element aux variable. "
-                        "Setting this to false will allow contribution to a single element's "
-                        "elemental value(s) from multiple boundary sides on the same element "
-                        "(example: when the restricted boundary exists on two or more sides "
-                        "of an element, such as at a corner of a mesh");
-
-  // This flag is set to true if the AuxKernelTempl is being used on a boundary
-  params.addPrivateParam<bool>("_on_boundary", false);
-
-  params.addRelationshipManager("GhostLowerDElems",
-                                Moose::RelationshipManagerType::GEOMETRIC |
-                                    Moose::RelationshipManagerType::ALGEBRAIC);
-
-  params.declareControllable("enable"); // allows Control to enable/disable this type of object
   params.registerBase("AuxKernel");
-  params.registerSystemAttributeName("AuxKernel");
-
   if (typeid(AuxKernelTempl<ComputeValueType>).name() == typeid(VectorAuxKernel).name())
     params.registerBase("VectorAuxKernel");
   if (typeid(AuxKernelTempl<ComputeValueType>).name() == typeid(ArrayAuxKernel).name())
     params.registerBase("ArrayAuxKernel");
+
   return params;
 }
 
 template <typename ComputeValueType>
 AuxKernelTempl<ComputeValueType>::AuxKernelTempl(const InputParameters & parameters)
-  : MooseObject(parameters),
+  : AuxKernelBase(parameters),
     MooseVariableInterface<ComputeValueType>(
         this,
         parameters.getCheckedPointerParam<SystemBase *>("_sys")
@@ -96,40 +53,11 @@ AuxKernelTempl<ComputeValueType>::AuxKernelTempl(const InputParameters & paramet
             : (std::is_same<RealVectorValue, ComputeValueType>::value
                    ? Moose::VarFieldType::VAR_FIELD_VECTOR
                    : Moose::VarFieldType::VAR_FIELD_ARRAY)),
-    BlockRestrictable(this),
-    BoundaryRestrictable(this, mooseVariableBase()->isNodal()),
-    SetupInterface(this),
-    CoupleableMooseVariableDependencyIntermediateInterface(this, mooseVariableBase()->isNodal()),
-    FunctionInterface(this),
-    UserObjectInterface(this),
-    TransientInterface(this),
-    MaterialPropertyInterface(this, blockIDs(), boundaryIDs()),
-    PostprocessorInterface(this),
-    DependencyResolverInterface(),
-    RandomInterface(parameters,
-                    *parameters.getCheckedPointerParam<FEProblemBase *>("_fe_problem_base"),
-                    parameters.get<THREAD_ID>("_tid"),
-                    mooseVariableBase()->isNodal()),
-    GeometricSearchInterface(this),
-    Restartable(this, "AuxKernels"),
-    MeshChangedInterface(parameters),
-    VectorPostprocessorInterface(this),
-    ElementIDInterface(this),
-    NonADFunctorInterface(this),
-    _check_boundary_restricted(getParam<bool>("check_boundary_restricted")),
-    _subproblem(*getCheckedPointerParam<SubProblem *>("_subproblem")),
-    _sys(*getCheckedPointerParam<SystemBase *>("_sys")),
-    _nl_sys(*getCheckedPointerParam<SystemBase *>("_nl_sys")),
-    _aux_sys(static_cast<AuxiliarySystem &>(_sys)),
-    _tid(parameters.get<THREAD_ID>("_tid")),
+
     _var(_aux_sys.getActualFieldVariable<ComputeValueType>(
         _tid, parameters.get<AuxVariableName>("variable"))),
     _nodal(_var.isNodal()),
     _u(_nodal ? _var.nodalValueArray() : _var.sln()),
-
-    _assembly(_subproblem.assembly(_tid, 0)),
-    _bnd(boundaryRestricted()),
-    _mesh(_subproblem.mesh()),
 
     _test(_bnd ? _var.phiFace() : _var.phi()),
     _q_point(_bnd ? _assembly.qPointsFace() : _assembly.qPoints()),
@@ -146,98 +74,8 @@ AuxKernelTempl<ComputeValueType>::AuxKernelTempl(const InputParameters & paramet
     _current_boundary_id(_assembly.currentBoundaryID()),
     _solution(_aux_sys.solution()),
 
-    _current_lower_d_elem(_assembly.lowerDElem()),
-    _coincident_lower_d_calc(_bnd && !isNodal() && _var.isLowerD())
+    _current_lower_d_elem(_assembly.lowerDElem())
 {
-  addMooseVariableDependency(&_var);
-  _supplied_vars.insert(parameters.get<AuxVariableName>("variable"));
-
-  if (_bnd && !isNodal() && !_coincident_lower_d_calc && _check_boundary_restricted)
-  {
-    // when the variable is elemental and this aux kernel operates on boundaries,
-    // we need to check that no elements are visited more than once through visiting
-    // all the sides on the boundaries
-    auto boundaries = _mesh.getMesh().get_boundary_info().build_side_list();
-    std::set<dof_id_type> elements;
-    for (const auto & t : boundaries)
-    {
-      if (hasBoundary(std::get<2>(t)))
-      {
-        const auto eid = std::get<0>(t);
-        const auto stat = elements.insert(eid);
-        if (!stat.second) // already existed in the set
-          mooseError(
-              "Boundary restricted auxiliary kernel '",
-              name(),
-              "' has element (id=",
-              eid,
-              ") connected with more than one boundary sides.\nTo skip this error check, "
-              "set 'check_boundary_restricted = false'.\nRefer to the AuxKernel "
-              "documentation on boundary restricted aux kernels for understanding this error.");
-      }
-    }
-  }
-
-  // Check for supported variable types
-  // Any 'nodal' family that actually has DoFs outside of nodes, or gradient dofs at nodes is
-  // not properly set by AuxKernelTempl::compute
-  // NOTE: We could add a few exceptions, lower order from certain unsupported families and on
-  //       certain element types only have value-DoFs on nodes
-  const auto type = _var.feType();
-  if (_var.isNodal() && !((type.family == LAGRANGE) || (type.order <= FIRST)))
-    paramError("variable",
-               "Variable family " + Moose::stringify(type.family) + " is not supported at order " +
-                   Moose::stringify(type.order) + " by the AuxKernel system.");
-}
-
-template <typename ComputeValueType>
-const std::set<std::string> &
-AuxKernelTempl<ComputeValueType>::getRequestedItems()
-{
-  return _depend_vars;
-}
-
-template <typename ComputeValueType>
-const std::set<std::string> &
-AuxKernelTempl<ComputeValueType>::getSuppliedItems()
-{
-  return _supplied_vars;
-}
-
-template <typename ComputeValueType>
-void
-AuxKernelTempl<ComputeValueType>::addUserObjectDependencyHelper(const UserObject & uo) const
-{
-  _depend_uo.insert(uo.name());
-  for (const auto & indirect_dependent : uo.getDependObjects())
-    _depend_uo.insert(indirect_dependent);
-}
-
-template <typename ComputeValueType>
-void
-AuxKernelTempl<ComputeValueType>::addPostprocessorDependencyHelper(
-    const PostprocessorName & name) const
-{
-  getUserObjectBaseByName(name); // getting the UO will call addUserObjectDependencyHelper()
-}
-
-template <typename ComputeValueType>
-void
-AuxKernelTempl<ComputeValueType>::addVectorPostprocessorDependencyHelper(
-    const VectorPostprocessorName & name) const
-{
-  getUserObjectBaseByName(name); // getting the UO will call addUserObjectDependencyHelper()
-}
-
-template <typename ComputeValueType>
-void
-AuxKernelTempl<ComputeValueType>::coupledCallback(const std::string & var_name, bool is_old) const
-{
-  if (!is_old)
-  {
-    const auto & var_names = getParam<std::vector<VariableName>>(var_name);
-    _depend_vars.insert(var_names.begin(), var_names.end());
-  }
 }
 
 template <typename ComputeValueType>
