+++ framework/src/interfaces/NonADFunctorInterface.C
@@ -19,3 +19,11 @@ NonADFunctorInterface::NonADFunctorInterface(const MooseObject * const moose_obj
   : FunctorInterface(moose_object)
 {
 }
+
+#ifdef MOOSE_KOKKOS_ENABLED
+NonADFunctorInterface::NonADFunctorInterface(const NonADFunctorInterface & object,
+                                             const Moose::Kokkos::FunctorCopy & key)
+  : FunctorInterface(object, key)
+{
+}
+#endif
